+++ src/execution/plugins/script-execution-tools/script-execution-tools.test.ts
@@ -3,6 +3,7 @@ import spaceTrim from 'spacetrim';
 import { PromptTemplatePipeline } from '../../../classes/PromptTemplatePipeline';
 import { promptTemplatePipelineStringToJson } from '../../../conversion/promptTemplatePipelineStringToJson';
 import { PromptTemplatePipelineString } from '../../../types/PromptTemplatePipelineString';
+import { assertsExecutionSuccessful } from '../../assertsExecutionSuccessful';
 import { createPtpExecutor } from '../../createPtpExecutor';
 import { MockedEchoNaturalExecutionTools } from '../natural-execution-tools/mocked/MockedEchoNaturalExecutionTools';
 import { CallbackInterfaceTools } from '../user-interface-execution-tools/callback/CallbackInterfaceTools';
@@ -64,5 +65,9 @@ describe('createPtpExecutor + executing scripts in ptp', () => {
         expect(ptpExecutor({}, () => {})).resolves.toMatchObject({
             errors: [new Error(`Parameter {thing} is not defined`)],
         });
+
+        expect(() => ptpExecutor({}, () => {}).then(assertsExecutionSuccessful)).rejects.toThrowError(
+            /Parameter \{thing\} is not defined/,
+        );
     });
 });